split off insertDomTag into util for
