+++ packages/app/obojobo-document-engine/src/scripts/common/util/insert-dom-tag.js
@@ -1,0 +1,10 @@
+const insertDomTag = (props, type) => {
+	const tag = document.createElement(type)
+	for(const i in props){
+		tag[i] = props[i]
+	}
+	const firstTag = document.getElementsByTagName(type)[0]
+	firstTag.parentNode.insertBefore(tag, firstTag)
+}
+
+export default insertDomTag
